# gateway/docs/swagger.yaml
basePath: /
definitions:
  errors.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  models.CreatePrivateChatRequest:
    properties:
      friend_id:
        type: string
    type: object
  models.CreatePrivateChatResponse:
    properties:
      chat_id:
        type: string
    type: object
  models.CreatePublicChatRequest:
    properties:
      name:
        type: string
      participant_id:
        items:
          type: string
        type: array
    type: object
  models.GetChatUsersResponse:
    properties:
      users:
        items:
          type: string
        type: array
    type: object
  models.GetMeChatsResponse:
    properties:
      chat_id:
        items:
          type: string
        type: array
    type: object
...
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.RegisterResponse:
    properties:
      id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API Gateway for the Messenger service
  title: Messenger API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Выполняет авторизацию пользователя и генерирует JWT токен. Принимает
        данные для входа и возвращает токен при успешной авторизации.
      parameters:
      - description: Данные для авторизации пользователей
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная авторизация, возвращает JWT токен
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Неверные данные для авторизации
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Не авторизован, неверные учетные данные
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Принимает данные для регистрации
        и возвращает ответ с информацией о созданном пользователе.
      parameters:
      - description: Данные для регистрации нового пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Неверные данные для регистрации
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Регистрация нового пользователя
      tags:
      - auth
  /chat:
    get:
      consumes:
      - application/json
      description: Retrieve all user chats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetMeChatsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get My Chats
      tags:
      - chats
  /chat/{chat_id}/users:
    get:
      consumes:
      - application/json
      description: получить пользователей чата
      parameters:
      - description: ID чата
        in: path
        name: chat_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetChatUsersResponse'
        default:
          description: ""
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: GetChatUsers
      tags:
      - chats
  /chat/private:
    post:
      consumes:
      - application/json
      description: Create a private chat between users
      parameters:
      - description: Private Chat Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreatePrivateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreatePrivateChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Private Chat
      tags:
      - chats
  /chat/public:
    post:
      consumes:
      - application/json
      description: Create a public group chat
      parameters:
      - description: Public Chat Data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreatePublicChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreatePrivateChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Public Chat
      tags:
      - chats
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"